feat: implementation get officers

+++ api/src/presentation/api/resources/officer/routes.py
@@ -1,3 +1,4 @@
+from typing import List
 from fastapi import APIRouter, Depends, status
 from fastapi.responses import JSONResponse
 from src.domain import OfficerBadgeExists, OfficerNotFound, OfficerLoginNotFound
@@ -9,7 +10,7 @@ from src.presentation.api.resources.officer.request_model import (
     OfficerLogin,
     OfficerUpdate,
 )
-from src.presentation.api.resources.officer.response_model import Token
+from src.presentation.api.resources.officer.response_model import Token, OfficerRead
 
 officer_router = APIRouter()
 
@@ -27,6 +28,32 @@ def officer_login(
             data=Token(access_token=access_token, refresh_token=refresh_token),
         )
     except (OfficerNotFound, OfficerLoginNotFound) as e:
+        return JSONResponse(
+            status_code=status.HTTP_404_NOT_FOUND,
+            content=ResponseModel(error=True, message=str(e), data={}).model_dump(),
+        )
+
+
+@officer_router.get("/officer", response_model=ResponseModel[List[OfficerRead]])
+def get_officers(
+    usecase=Depends(officer_usecase_stub),
+    current_officer=Depends(get_current_officer_stub),
+):
+    try:
+        if RoleEnum(current_officer.role) != RoleEnum.ADMIN:
+            return JSONResponse(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                content=ResponseModel(
+                    error=True, message="Unauthorized", data={}
+                ).model_dump(),
+            )
+        officers = usecase.get_officers()
+        return ResponseModel(
+            error=False,
+            message="Officers retrieved successfully",
+            data=officers,
+        )
+    except OfficerNotFound as e:
         return JSONResponse(
             status_code=status.HTTP_404_NOT_FOUND,
             content=ResponseModel(error=True, message=str(e), data={}).model_dump(),
@@ -68,6 +95,13 @@ def update_officer(
     current_officer=Depends(get_current_officer_stub),
 ):
     try:
+        if RoleEnum(current_officer.role) != RoleEnum.ADMIN:
+            return JSONResponse(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                content=ResponseModel(
+                    error=True, message="Unauthorized", data={}
+                ).model_dump(),
+            )
         badge = current_officer.get("badge")
         usecase.update_officer(id=id, badge=badge, update_officer=data)
         return ResponseModel(
@@ -89,6 +123,13 @@ def delete_officer(
     current_officer=Depends(get_current_officer_stub),
 ):
     try:
+        if RoleEnum(current_officer.role) != RoleEnum.ADMIN:
+            return JSONResponse(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                content=ResponseModel(
+                    error=True, message="Unauthorized", data={}
+                ).model_dump(),
+            )
         badge = current_officer.get("badge")
         usecase.delete_officer(id=id, badge=badge)
         return ResponseModel(
